+++ test/codegen/gql2ts.test.ts
@@ -1,3 +1,4 @@
+import * as fs from "fs";
 import got from "got";
 import { gql2ts } from "../../src";
 import { requireExampleServer, TestLogger } from "../util";
@@ -66,6 +67,39 @@ describe("The gql2ts function", () => {
       server: true,
       noExit: true,
       stitching: true,
+      logger,
+    });
+
+    expect(logger.errors).toEqual([]);
+  });
+
+  it("should write an introspection fallback file if the file option is set", async () => {
+    const logger = new TestLogger();
+    const file = "test/codegen/introspection.json";
+
+    await fs.promises.rm(file).catch(() => 0);
+
+    await gql2ts({
+      root: "examples/client",
+      introspection: {
+        url: "http://localhost:4999/graphql",
+        file,
+      },
+      client: true,
+      noExit: true,
+      logger,
+    });
+
+    await fs.promises.readFile(file, "utf-8");
+
+    await gql2ts({
+      root: "examples/client",
+      introspection: {
+        url: "http://localhost:4999/use/fallback/instead",
+        file,
+      },
+      client: true,
+      noExit: true,
       logger,
     });